7119における発熱等相談件数，救急医療の東京ルールの適用件数: Rename components/MixedBarAndLineChart.vue to components/index/_shared/MixedBarAndLineChart.vue
